+++ spec/messages.js
@@ -2010,7 +2010,7 @@ describe("A XEP-0333 Chat Marker", function () {
         spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
         spyOn(view.model, 'sendMarker').and.callThrough();
         _converse.connection._dataRecv(mock.createRequest(stanza));
-        await u.waitUntil(() => view.model.sendMarker.calls.count() === 1);
+        await u.waitUntil(() => view.model.sendMarker.calls.count() === 2);
         expect(Strophe.serialize(sent_stanzas[0])).toBe(
             `<message from="romeo@montague.lit/orchard" `+
                     `id="${sent_stanzas[0].nodeTree.getAttribute('id')}" `+
@@ -2044,7 +2044,7 @@ describe("A XEP-0333 Chat Marker", function () {
             .map(s => _.isElement(s) ? s : s.nodeTree)
             .filter(e => e.nodeName === 'message');
 
-        expect(sent_messages.length).toBe(1);
+        expect(sent_messages.length).toBe(2);
         expect(Strophe.serialize(sent_messages[0])).toBe(
             `<message id="${sent_messages[0].getAttribute('id')}" to="${contact_jid}" type="chat" xmlns="jabber:client">`+
                 `<active xmlns="http://jabber.org/protocol/chatstates"/>`+